CI: Add release workflow
This reuses the build stage action as well as the build-container
job from the testing workflow. Instead of test installing the package
and pruning old images however, this workflow will publish
a new release with the package and package checksum as assets.

The testing workflow has also been changed to only apply to push
events on branches.