CI: Avoid storing unnessarry data
After the container got built and is stored in
the local registry, there is no need to keep the build
source archive as well as all depencies around.

Additionally, its not needed to keep the imported container
image around in a build stage.

This fixes runtime errors in CI due to storage restrictions
of the GitHub Workers.